fix: typo

+++ javascript/lucky-numbers/test.js
@@ -4,7 +4,7 @@ import { luckyNumber, twoSum } from "./lucky-numbers.js";
 
 function testFunctions(){
     console.log(twoSum([1, 2, 4, 0, 3, 5, 2, 9], [3, 2, 4, 8, 1, 5, 4, 1, 8]));
-    console.logf(luckyNumber(15651));
+    console.log(luckyNumber(15651));
 }
 
 testFunctions();
